Added ModelState check in controllers

+++ projects/TemplateSolution/src/TemplateProject.Web/frontend/src/app/main/pages/todo/index/todo-index.component.ts
@@ -3,6 +3,7 @@ import { Router } from '@angular/router';
 
 import { TodoItemsService } from 'src/app/main/services/todo/todo-items.service';
 import { DialogService } from 'src/app/shared/dialogs/dialog.service';
+import { ConfirmResult } from 'src/app/shared/dialogs/confirm/confirm-result';
 
 @Component({
     selector: 'app-todo-index',
@@ -32,12 +33,20 @@ export class TodoIndexComponent implements OnInit {
     }
 
     onDeleteButtonClicked(item: any): boolean {
-        this._todoItemsService.delete(item.id)
-            .subscribe(() => {
-                this.items.splice(this.items.indexOf(item), 1);
+        this._dialogService.showConfirmAsync({ title: `Are you sure to delete user '${item.name}'?` })
+            .then(result => {
+                if (result != ConfirmResult.Yes)
+                    return;
 
-                this._dialogService.showSuccess();
-            }, error => this._dialogService.showErrorAsync({ text: error.message }));
+                this._dialogService.showExecuting();
+
+                this._todoItemsService.delete(item.id)
+                    .subscribe(() => {
+                        this.items.splice(this.items.indexOf(item), 1);
+
+                        this._dialogService.showSuccess();
+                    }, () => this._dialogService.showErrorAsync({ text: 'A error occured while deleting the item.' }));
+            });
 
         return false;
     }
